test: 💍 fragment_cobmination: fixed bug with conn > 2 and tests
Only the last IR was used for computing the connection points (cp1,
cp2). This was fixed using a list that gets incremented. Might not be
the best of the best implementations (sets could have worked right away
as well), but this is working. All tests have been updated to check for
the fc. There is still a need for examples for cases cbl, ctl and col.