feat: 🎸 duplicate: option for returning the df of filtered mols
BREAKING CHANGE: 🧨 A bug was found in the logs: the imported module log is no longer
displayed, so my nice --log DEBUG argument is now useless. I will have
to check why exactly since it works fine in the juypter notebooks. To
work around this problem, I now return a df and write an extra csv file
in the log subfolder. This will prove useful for parsing results anyway
so why not.