feat: 🎸 bin/run_protocol: helper file for smarter reruns
Write a small json file with the chunksize, counts of molecules and
corresponding number of chunks. It is not super optimized, because if
one changes the number of chunks, then the count of molecules has to
recomputed. But frankly speaking, I will not use this feature super
often and I am a bit short of time.