fix: 🐛 load/save: now use smiles instead of pickle for mols
Ok This was a very weird one. I have some molecules going on missing
during standardization. There is no reason whatsoever for this, it just
seems to happen. If I run the molecules locally, it works. Since I have
a lot of warnings about pickle version incompatibility, I thought maybe,
I could switch from pickle to smiles storage instead. I really dislike
this, because of work overload and also impossibility to keep nice
coordinates. But for now it is what I do. For this reason I enabled onl
rdDepictor as depiction method in the smk protocols.